progreso

+++ seguridad/usuario.service.js
@@ -4,7 +4,7 @@ var config = require('config.json');
 var jwt = require('jsonwebtoken');
 var bcrypt = require('bcryptjs');
 var db = require('_helpers/db');
-var User = db.Usuario;
+var User = db['Usuario'];
 var randT = require('rand-token');
 
 
@@ -16,7 +16,9 @@ module.exports = {
     authenticate,
     getAll,
     getById,
+    getByProp,
     create,
+    createMany,
     update,
     logout,
     delete: _delete,
@@ -33,7 +35,7 @@ async function isLoggedIn({
             try {
                 jwt.verify(user.accesToken, config.secret);
             } catch (error) {
-                // console.log(error.name, ' ', error.message);
+                // ////console.log(error.name, ' ', error.message);
                 if (error.name === 'TokenExpiredError' && user.loggedIn) {
                     return (false);
                 }
@@ -45,7 +47,7 @@ async function isLoggedIn({
 }
 
 async function gNewTokenAcces(usernameparam, tokenRefresh) {
-    // console.log('Data in refreshTokens: ', refreshTokens);
+    // ////console.log('Data in refreshTokens: ', refreshTokens);
     // await db.dropCollection('Colection');
     // await db.dropCollection('Subcategoria');
     refreshTokens[tokenRefresh] = usernameparam;
@@ -53,17 +55,17 @@ async function gNewTokenAcces(usernameparam, tokenRefresh) {
         var user = await User.findOne({
             username: usernameparam
         });
-        // console.log(user);
+        // ////console.log(user);
         if (user) {
             user.accesToken = jwt.sign({
-                    name: user.username,
-                    sub: user.id,
-                    rol: user.rol
-                },
+                name: user.username,
+                sub: user.id,
+                rol: user.rol
+            },
                 config.secret, {
-                    expiresIn: 60
-                });
-            // console.log(user.accesToken);
+                expiresIn: 60
+            });
+            // ////console.log(user.accesToken);
             await user.save();
             return (user.accesToken);
         } else {
@@ -79,33 +81,40 @@ async function authenticate({
     password
 }) {
     const user = await User.findOne({
-        username:username
+        username: username
     });
-    if(bcrypt.compareSync(password,user.password)){
-    user.loggedIn = true;
-    const {
-        hash,
-        ...userWithoutHash
-    } = user.toObject();
-    const token = jwt.sign({
-            name: username,
-            sub: user.id,
-            rol: ''
-        },
-        config.secret, {
-            expiresIn: 60
-        });
-    
-    user.accesToken = token;
-    const tokenRefresh = randT.generate(16);
-    refreshTokens[tokenRefresh] = username;
-    await user.save();
-    return {
-        ...userWithoutHash,
-        token,
-        tokenRefresh
-    };
-}
+    if (user.password) {
+        if (bcrypt.compareSync(password, user.password)) {
+            user.loggedIn = true;
+            const {
+                hash,
+                ...userWithoutHash
+            } = user.toObject();
+            const token = jwt.sign({
+                name: username,
+                sub: user.id,
+                rol: ''
+            },
+                config.secret, {
+                expiresIn: 60
+            });
+
+            user.accesToken = token;
+            const tokenRefresh = randT.generate(16);
+            refreshTokens[tokenRefresh] = username;
+            await user.save();
+            return {
+                ...userWithoutHash,
+                token,
+                tokenRefresh
+            };
+        }
+    }else{
+        return {
+            result:'FIRST_LOGIN',
+            data:user
+        }
+    }
 
     // if (user && password === user.password) {
     // }
@@ -115,24 +124,72 @@ async function getAll() {
     return await User.find().select('-hash');
 }
 
-async function getById({...query}) {
+async function getByProp({ ...query }) {
     return await User.findOne(query).select('-hash');
+}
+async function getById(id) {
+    return await User.findById(id);
+
 }
 
 async function create(userParam) {
     if (await User.findOne({
-            username: userParam.username
-        })) {
+        username: userParam.username
+    })) {
         throw 'Username"' + userParam.username + '" is already taken';
     }
     const user = new User(userParam);
+    if (userParam.password) {
+        user.password = bcrypt.hashSync(userParam.password, 10);
+    }
     await user.save((error) => {
         if (error) {
-            console.log(error);
-           return error;
+            ////console.log(error);
+            return error;
         }
     });
     return user;
+}
+
+async function createMany(usersParams) {
+
+    const indices = usersParams.map((current) => current.username);
+    // ////console.log(indices);
+    let users = [];
+    let exist = [];
+    let errors = [];
+    let result = new Object();
+    await User.find({
+        username: { $in: indices }
+    }, (err, arr) => {
+        if (arr) {
+            // ////console.log(arr);
+            exist = arr.map((current) => current.username);
+        }
+    });
+    const add = [];
+    usersParams.forEach((current) => {
+        if (!exist.includes(current.username) && (current.username.length <= 8)) {
+            let u = new User(current);
+            add.push(u);
+        }
+    });
+    if (add.length > 0) {
+        await User.insertMany(add, (error, succ) => {
+            if (succ) {
+                users.push(succ);
+            }
+            if (error) {
+                errors.push(error);
+            }
+        })
+    }
+    result['result'] = users;
+    result['exist'] = exist;
+    result['errors'] = errors;
+    return result;
+
+
 }
 
 async function update(id, userParam) {
@@ -141,8 +198,8 @@ async function update(id, userParam) {
     if (!user) throw 'User not found';
 
     if (user.username !== userParam.username && await User.findOne({
-            username: userParam.username
-        })) {
+        username: userParam.username
+    })) {
         throw 'username "' + userParam.username + '" is already taken';
     }
 
@@ -158,7 +215,7 @@ async function logout(id, userParam) {
 
     var user = await User.findById(id);
     if (!user) {
-        console.log('User not found');
+        ////console.log('User not found');
         throw 'User not found';
     }
     user.loggedIn = false;
@@ -167,5 +224,7 @@ async function logout(id, userParam) {
 }
 
 async function _delete(id) {
-    await User.findByIdAndRemove(id);
+    if(id){
+        await User.findByIdAndRemove(id);
+    }
 }
